Update to Kotlin 1.2.71 + Coroutines 0.30.0

+++ base/src/main/java/app/tivi/extensions/RetrofitExtensions.kt
@@ -32,7 +32,7 @@ fun <T> Response<T>.bodyOrThrow(): T {
 fun <T> Response<T>.toException() = HttpException(this)
 
 suspend inline fun <T> Call<T>.executeWithRetry(
-    firstDelay: Int = 100,
+    firstDelay: Long = 100,
     maxAttempts: Int = 3,
     shouldRetry: (Exception) -> Boolean = ::defaultShouldRetry
 ): Response<T> {
@@ -59,7 +59,7 @@ suspend inline fun <T> Call<T>.executeWithRetry(
 }
 
 suspend inline fun <T> Call<T>.fetchBodyWithRetry(
-    firstDelay: Int = 100,
+    firstDelay: Long = 100,
     maxAttempts: Int = 3,
     shouldRetry: (Exception) -> Boolean = ::defaultShouldRetry
 ) = executeWithRetry(firstDelay, maxAttempts, shouldRetry).bodyOrThrow()
